pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "clipwarden"
version = "0.1.0"
description = "ClipWarden: Windows background tool that detects clipboard hijacking of cryptocurrency addresses."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Ethan Tharp", email = "ethan@ethantharp.dev" }]
keywords = ["security", "cryptocurrency", "clipboard", "windows", "malware", "clipper"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Win32 (MS Windows)",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: Microsoft :: Windows :: Windows 10",
  "Operating System :: Microsoft :: Windows :: Windows 11",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
]

# Runtime deps are declared here for tooling that reads pyproject metadata.
# The authoritative pinned + hashed list lives in requirements.txt, generated
# from requirements.in by pip-compile. Install prod with:
#   pip install --require-hashes -r requirements.txt
dependencies = [
  "pywin32",
  "pystray",
  "Pillow",
  "winotify",
  "base58",
  "bech32",
  "pynacl",
]

[project.urls]
Homepage = "https://ethantharp.dev"
Source = "https://github.com/ItsZardos/clipwarden"
Issues = "https://github.com/ItsZardos/clipwarden/issues"

[project.scripts]
clipwarden = "clipwarden.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/clipwarden"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests", "tools"]

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "F",    # pyflakes
  "W",    # pycodestyle warnings
  "I",    # isort
  "B",    # bugbear
  "UP",   # pyupgrade
  "SIM",  # simplify
]
ignore = [
  "E501",  # line length handled by formatter, not lint
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
filterwarnings = [
  "error",
  # pywin32 emits a DeprecationWarning about pkg_resources on import.
  # Not ours to fix, don't fail the suite over it.
  "ignore::DeprecationWarning:pywintypes",
]

[tool.coverage.run]
source = ["clipwarden"]
branch = true
# __main__ is the CLI stub covered by a subprocess smoke test; coverage
# does not see it because that test shells out.
omit = ["src/clipwarden/__main__.py"]

[tool.coverage.report]
exclude_also = [
  "if TYPE_CHECKING:",
  "raise NotImplementedError",
]
show_missing = true
